installs nvm and rollsback to node 8.13 to try and resolve the functions.config() sometimes undefined issue
